時間データ処理関数きりだし
